+++ libreyolo/export/exporter.py
@@ -856,6 +856,12 @@ class BaseExporter(ABC):
             nn_model = LWDETRExportWrapper(nn_model).to(device)
             nn_model.eval()
             dfine_wrapped = True
+        elif family == "faster_rcnn":
+            from ..models.faster_rcnn.nn import FasterRCNNExportWrapper
+
+            nn_model = FasterRCNNExportWrapper(nn_model).to(device)
+            nn_model.eval()
+            dfine_wrapped = True
         elif family == "deformable_detr":
             from ..models.deformable_detr.nn import DeformableDETRExportWrapper